shadow_shmem
Module raw_syscall

A simple interface to make system calls without relying on libc or nix, which are not no-std.

The goal of this module is to be as no-std as possible to be compatible with Shadow’s shim preload library.

Public functions for system calls are named according to their standard names found in man(2). The public function accept have parameters with types that are natural for Rust. For a list of syscalls, a good reference is: https://blog.rchapman.org/posts/Linux_System_Call_Table_for_x86_64/

Constants

S_IRGRP

S_IRUSR

S_IWGRP

S_IWUSR

Functions

clock_monotonic_gettime

close

fsync

ftruncate

getpid

gettid

kill

mmap⚠

Safety

munmap

open⚠

Safety

tgkill

unlink⚠

Safety

write⚠

Safety